pyproject.toml
[project]
name = "autoresearch"
version = "0.1.0"
description = "A high-performance Python application template"
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.12"
dependencies = [
    "click>=8.3.1",
    "httpx>=0.28.1",
    "mcp>=1.26.0",
    "msgspec>=0.20.0",
    "orjson>=3.11.7",
    "pyyaml>=6.0.3",
    "structlog>=25.5.0",
]

[project.scripts]
autoresearch = "autoresearch.cli:main"
autoresearch-mcp-server = "autoresearch.adapters.mcp_server:main"

[build-system]
requires = ["uv_build>=0.10.7,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.1",
    "pytest-cov>=7.1.0",
    "pytest-asyncio>=1.3.0",
    "pytest-benchmark>=5.2.3",
    "ruff>=0.15.5",
    "ty>=0.0.21",
    "behave>=1.3.3",
    "hypothesis>=6.151.9",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "benchmark: marks benchmark tests",
]
asyncio_mode = "auto"
